requirements.txt
pyOpenSSL
asn1crypto
certvalidator
tabulate
prompt_toolkit
tqdm
